refactor: Tests refactoring
